% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available.costmethods}
\alias{available.costmethods}
\title{A list of available related cost correction methods by country}
\format{
list
}
\source{
PAID4 tool
}
\usage{
available.costmethods
}
\description{
If a country conducts a comprehensive cost-of-illness study and it has been
included in PAID, then the default cost method is "dscosts", else "totcosts".
The data shows which method is available per country. This dataset shows the
possible options by country for the \code{cost.method} option under the \code{paid}
function.
}
\keyword{datasets}
