% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.discount}
\alias{f.discount}
\title{Generate discount rate vector}
\usage{
f.discount(x)
}
\arguments{
\item{x}{either a single numerical value (in percent) or a numerical named
vector}
}
\value{
a vector of discount rates
}
\description{
This function enables users to adjust the discount percentage depending on
the time horizon. Some countries have a different discount rate depending on
the time horizon. For example, France's discount rate is 2.5\% for the first
29 years, and 1.5\% after 30 or more years.
}
