% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.draw}
\alias{f.draw}
\title{Draw ratios}
\usage{
f.draw(par1, par2, N)
}
\arguments{
\item{par1}{the log-mean of the ratio distribution}

\item{par2}{the standard deviation of the ratio distribution}

\item{N}{the number of draws. Taken from the main function.}
}
\value{
an array
}
\description{
Draw ratios
}
