% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.left}
\alias{f.left}
\title{Extract the first characters from text}
\usage{
f.left(text, num_char)
}
\arguments{
\item{text}{a character vector}

\item{num_char}{an integer or an integer vector}
}
\value{
a character vector
}
\description{
This is the equivalent of Excel LEFT function.
}
\examples{
f.left("This text", 3)
f.left(c("This text","That text"), c(3,4))
}
