% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.lowhighages}
\alias{f.lowhighages}
\title{Helper function to set the same value at low and high ages}
\usage{
f.lowhighages(d, a1 = 51, a2 = 96)
}
\arguments{
\item{d}{A numeric vector of length 202. First 101 are for men.}

\item{a1}{An integer defining the first location of the lowest age.}

\item{a2}{An integer defining the first location of the highest age.}
}
\value{
A numeric vector
}
\description{
Helper function to set the same value at low and high ages
}
