% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.mid}
\alias{f.mid}
\title{Extract from the middle of text}
\usage{
f.mid(text, start_num, num_char)
}
\arguments{
\item{text}{a character vector}

\item{start_num}{an integer denoting the start}

\item{num_char}{an integer denoting how many characters to extract}
}
\value{
a character vector
}
\description{
This is the equivalent of Excel MID function.
}
\examples{
f.mid("This text", 6,2)
}
