% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.replacespecial}
\alias{f.replacespecial}
\title{Ensures no negative values for mean (low, high)}
\usage{
f.replacespecial(x, repif = NULL, repwith = NULL)
}
\arguments{
\item{x}{a matrix with at least three columns containing the names "mean","lower", and "upper".}

\item{repif}{replace values of \code{x} if they are equal or less than \code{repif}}

\item{repwith}{replace values of \code{x} if they are equal or less than \code{repif} with \code{repwith}}
}
\value{
a matrix
}
\description{
Substitutes with the latest non-negative value but also ensures the accompanying confidence interval remains correct.
}
