% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.right}
\alias{f.right}
\title{Extract from the right of text}
\usage{
f.right(text, num_char)
}
\arguments{
\item{text}{a character vector}

\item{num_char}{an integer or an integer vector}
}
\value{
a character vector
}
\description{
This is the equivalent of RIGHT MID function.
}
