% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.smoothAC}
\alias{f.smoothAC}
\title{Smooths out any vector into 1-year age groups}
\usage{
f.smoothAC(dep, inc = 1, lastsexage)
}
\arguments{
\item{dep}{numerical vector where the first N values are for men and the next N are for women.}

\item{inc}{ensures a small increment is added to any value that is a perfect 0. This ensures correct log transformation.}

\item{lastsexage}{location within \code{dep} vector of the last age group for men.}
}
\value{
numerical vector of length 202
}
\description{
This package assumes throughout that data for men and women are stacked and are all thus sized 2*N, where N is the length of the vector for men.
A Generalized additive model with P-splines is used for the smoothing.
}
