% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{f.sumdiag}
\alias{f.sumdiag}
\title{Sums the diagonal of the decedent costs files}
\usage{
f.sumdiag(x, from, to, disc_vector)
}
\arguments{
\item{x}{a matrix or data.frame with spending at different ages as rows and spending by time-to-death as columns - 5 columns total.}

\item{from}{an integer showing location of the starting age. Ages start from 0 to 100, hence age 50 will have location 51 for example.}

\item{to}{an integer showing location of the end age.}

\item{disc_vector}{numerical vector of the discount needed to be applied to costs. Depends of user-chosen discount percentage and cycle length.}
}
\value{
a numerical vector of spending in the last 5 years of life given age.
}
\description{
Produces a vector of the spending within the last 5 years of life.
}
