% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{getSCDC}
\alias{getSCDC}
\title{Disaggregate per-capita spending}
\usage{
getSCDC(country, providers, coi.year, PSA, psa.N)
}
\arguments{
\item{country}{a string}

\item{providers}{a string vector}

\item{coi.year}{year as taken from main function.}

\item{PSA}{logical. Taken from main function.}

\item{psa.N}{number of draws. Taken from main function.}
}
\value{
a list
}
\description{
Attributes costs to survivors and decedents.
}
