% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping}
\alias{mapping}
\title{A list of diseases that can be excluded from the estimates}
\format{
list of data frames for each country
}
\source{
PAID 4.0 tool
}
\usage{
mapping
}
\description{
Only important for countries that have a cost of illness study - Netherlands,
Germany, and France. \code{paid4::mapping[[country]][[coi.year]]} displays a data frame
listing all diseases that comprise the cost estimates and show what their
corresponding groups are. To be used as guidance for \code{related.diseases}
option of the \code{paid} function.
}
\keyword{datasets}
