% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paid4-package.R
\docType{package}
\name{paid4-package}
\alias{paid4}
\alias{paid4-package}
\title{paid4: Practical Application to Include future Disease costs (PAID) 4.0}
\description{
PAID 4.0 is a toolkit that enables researchers to incorporate future unrelated medical costs and future non-medical costs in their economic evaluations. Users can specify various aspects of their intervention to then receive case-specific future unrelated medical costs and future non-medical costs.
}
\author{
\strong{Maintainer}: Elena Milkovska \email{milkovska@eshpm.eur.nl}

Authors:
\itemize{
  \item Pieter van Baal \email{vanbaal@eshpm.eur.nl}
}

}
\keyword{internal}
