% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderCohort}
\alias{renderCohort}
\title{Combines lifetime spending with survival data}
\usage{
renderCohort(survdata, costlist, pmen, cycle_length, start_age)
}
\arguments{
\item{survdata}{Takes the same \code{survdata} as the main function pai4::paid}

\item{costlist}{Takes the \code{lhce} output from renderLHCE}

\item{pmen}{Takes the same \code{pmen} as the main function pai4::paid}

\item{cycle_length}{Takes the same \code{cycle_length} as the main function pai4::paid}

\item{start_age}{Takes the same \code{start_age} as the main function pai4::paid}
}
\value{
a list
}
\description{
Combines lifetime spending with survival data
}
