% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderLHCE}
\alias{renderLHCE}
\title{Generate unrelated lifetime healthcare spending
Takes the survivor and decedent costs by age, removes any spending if specified by user, and combines into lifetime spending.}
\usage{
renderLHCE(
  country,
  user.providers,
  user.diseases,
  disc_percentage,
  user.uploaded,
  coi.year
)
}
\arguments{
\item{country}{Takes the same \code{country} as the main function pai4::paid}

\item{user.providers}{a character vector. Takes the same \code{user.providers} as the main function pai4::paid}

\item{user.diseases}{a data.frame of related diseases. See \code{related.diseases}in the main function pai4::paid}

\item{disc_percentage}{Takes the same \code{disc_percentage} as the main function pai4::paid}

\item{user.uploaded}{Takes the same \code{user.uploaded} as the main function pai4::paid}
}
\value{
list
}
\description{
Generate unrelated lifetime healthcare spending
Takes the survivor and decedent costs by age, removes any spending if specified by user, and combines into lifetime spending.
}
