% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{renderUnrelSCDC}
\alias{renderUnrelSCDC}
\title{Generates unrelated survivor and decedent costs
Adjusts spending by time to death with the probability to NOT die of either related diseases conditional on death at age a. Then sums op the spending across selected providers.}
\usage{
renderUnrelSCDC(country, user.providers, user.diseases, coi.year)
}
\arguments{
\item{country}{Takes the same \code{country} as the main function pai4::paid}

\item{user.providers}{a character vector. Takes the same \code{user.providers} as the main function pai4::paid}

\item{user.diseases}{a data.frame of related diseases. See \code{related.diseases}in the main function pai4::paid}
}
\description{
Generates unrelated survivor and decedent costs
Adjusts spending by time to death with the probability to NOT die of either related diseases conditional on death at age a. Then sums op the spending across selected providers.
}
