% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available.providers}
\alias{available.providers}
\title{A list of available providers by country}
\format{
list
}
\source{
PAID 4.0 tool
}
\usage{
available.providers
}
\description{
Each country has its own set of healthcare providers that comprise per-capita
healthcare spending. This file gives a list of available providers by country
to guide what can be placed in the \code{providers} option of the \code{paid} function.
}
\keyword{datasets}
