% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{paid}
\alias{paid}
\title{Produce unrelated medical spending}
\usage{
paid(
  country = c("Netherlands", "Germany", "France", "Greece", "Spain", "United Kingdom"),
  providers = "ALL",
  discount_perc = NULL,
  related.diseases = NULL,
  cost.method = NULL,
  related.costs = NULL,
  survdata = NULL,
  pmen = 0.5,
  cycle_length = 1,
  start_age = 0,
  coi.year = NULL,
  PSA = FALSE,
  psa.N = NULL,
  psa.seed = NULL
)
}
\arguments{
\item{country}{Current options are "Netherlands", "Germany","France",
"Greece", "Spain", and "United Kingdom".}

\item{providers}{By default all providers within a country are taken. Review
available providers by country \code{paid4::available.providers}. For example,
setting this to c("HC","LTC") will only show the sum of costs from
inpatient hospital and long-term care spending.}

\item{discount_perc}{either a single numerical value (in percent, e.g. 3) or
a numerical named vector (e.g., c('1' = 2.5, '30' = 1.5)). This will be the
percentage used to discount lifetime healthcare costs. The default is
assigned as per the country-specific HTA guidelines.}

\item{related.diseases}{an integer vector or \code{NULL}. Applicable for the
Netherlands, Germany, and France only. The default, \code{NULL}, marks all
diseases as unrelated. To mark a disease as related and exclude its costs,
provide here its corresponding "Group in PAID" number from
\code{paid4::mapping}. Example: \code{related.diseases = c(14,25)} for the
Netherlands would mark "breast cancer" and "other cancers" as related
diseases and exclude their costs.}

\item{cost.method}{\code{NULL} (default) or a character vector. Defines how costs
of related diseases are excluded - from cost of illness (option "dscosts")
or from own data "totcosts". The default for Netherlands, Germany, and
France is "dscosts", "totcosts" for the rest. Review options
\code{paid4::available.costmethods}.}

\item{related.costs}{\code{NULL} (default), numeric vector (see
\code{paid4::related.costs.ac}) or list (see \code{paid4::related.costs.scdc}). This
excludes any (additional) user-provided related spending by age, as needed.
This can be relevant for countries without cost of illness studies, for
example, to exclude per-capita costs of cancer from the total per-capita
survivor and decedent costs. Fill in the provided templates under
\code{related.costs.ac} or \code{related.costs.scdc} and pass it on as an argument
here retaining the same structure.}

\item{survdata}{By default \code{NULL}. Else takes a data.frame or matrix with
two columns. The first column must give overall survival of the
intervention group, and the second - of the control.}

\item{pmen}{Only relevant if \code{survdata} is provided. Proportion of men in the
cohort. Default is 0.5.}

\item{cycle_length}{Only relevant if \code{survdata} is provided. A numeric value
indicating the cycle length in years. The default is 1, meaning 1 year. For
example, a cycle length of 3 weeks would be 3/52 = 0.05769231.}

\item{start_age}{Only relevant if \code{survdata} is provided. Age at the start
of the simulation. Default is 0.}

\item{coi.year}{numerical value, denoting which Cost of Illness study year is
to be used for calculating lifetime healthcare spending. The default
differs by country, but it is always the latest COI year. Currently, the
default for the Netherlands is 2019, Germany - 2020, and France - 2019.}

\item{PSA}{logical. If \code{TRUE} ratios will be drawn 1000 times and the resulting
lifetime medical spending will be presented as a distribution.}

\item{psa.N}{integer. Defines how many draws in case of PSA. Default is 1000.}

\item{psa.seed}{integer. If PSA, by default, a random seed is generated and
reported back in the end results. User can also provide their own seed here.}
}
\value{
If no \code{survdata} is provided, the function returns a list containing:
\enumerate{
\item lhce: full age matrices of discounted but not half-cycle corrected
lifetime unrelated HCE, for men and women separately. The result comes back
as a named array of size \verb{[202,101,3]}. The first 101 rows are for men, and
102:202 are for women. Rows denote start age, while columns are the age at
death. The third dimension of the array is also named and returns either
the mean, lower of upper range of the estimate. For example, sub-setting
the \code{lhce} array as \verb{[1,,]} will return a 101x3 matrix of lifetime medical
spending since age 0 for men with its mean, lower and upper range.
\item sc: non-discounted survivor per-capita costs by age (TTD>4).
\item dc: non-discounted decedent costs (TTD<=4) where TTD=0 means death at
the same age as start.
\item Source year used for costs and presenting prices.
}

If \code{survdata} is provided, the function returns a list of
\enumerate{
\item output: Survival times discounted, half-cycle corrected, sex-weighted
unrelated costs;
\item wcosts:  Discounted, half-cycle corrected, sex-weighted unrelated costs;
\item uwcosts: Discounted, half-cycle corrected unrelated costs for men &
women separately in the pre-specified cycle length.
\item Source year used for costs and presenting prices.
}
}
\description{
Cost-effectiveness analysis (CEA) is used in many countries to evaluate
whether new drugs and technologies provide sufficient value for money as
compared to existing ones. In CEA, costs and benefits are estimated over a
patient’s lifetime in various scenarios. While lifetime medical spending
associated with the treated disease(s) is routinely included in CEA, the
inclusion of spending related to other diseases in life years gained is less
common practice (e.g. costs of treating dementia in life-years gained after a
successful heart transplant). These costs are referred to as future unrelated
medical costs or indirect medical costs (IMC). The current function produces
estimates of IMC for different sets of healthcare providers and comprising
diseases.
}
